+++ typatro/__main__.py
@@ -1,6 +1,6 @@
 import click
 
-PKG_VERSION = "1.0.2"
+PKG_VERSION = "1.0.3"
 
 
 @click.group(